requirements-cpu.txt
cryptography>=3.4.8
pyjwt>=2.1.0
waitress>=2.0.0
flask>=2.0.1
flask_cors>=3.0.10
werkzeug>=2.0.1
yapf>=0.31.0
pytest>=6.2.5
setuptools>=44.0.0
pyyaml>=5.4.1
Pillow>=8.4.0
numpy>=1.18.5
tqdm>=4.64.0
spacy>=3.3.1
tika>=1.24
pyvis>=0.2.1
ampligraph>=1.4.0
tensorflow==1.15.5